+++ .idea/workspace.xml
@@ -4,21 +4,15 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="b36bb03d-2f24-4880-8faa-28b42ad81b94" name="Changes" comment="">
-      <change afterPath="$PROJECT_DIR$/.gitignore" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/.idea/encodings.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/.idea/jpa-buddy.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/.idea/misc.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/.idea/uiDesigner.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/.idea/vcs.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/pom.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/abstractclasses/Circle.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/abstractclasses/Ellipse.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/abstractclasses/EqualTriangle.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/abstractclasses/RegularTriangle.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/abstractclasses/Shape.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/test/java/com/keyin/qap4/abstractclasses/Demo.java" afterDir="false" />
+    <list default="true" id="b36bb03d-2f24-4880-8faa-28b42ad81b94" name="Changes" comment="Flesh of the program--problem 1">
+      <change afterPath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/Scalable.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/abstractclasses/Circle.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/Circle.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/abstractclasses/Ellipse.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/Ellipse.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/abstractclasses/EqualTriangle.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/EqualTriangle.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/abstractclasses/RegularTriangle.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/RegularTriangle.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/abstractclasses/Shape.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/keyin/qap4/Shape.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/java/com/keyin/qap4/abstractclasses/Demo.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/com/keyin/qap4/Demo.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -34,6 +28,11 @@
     </option>
   </component>
   <component name="Git.Settings">
+    <option name="RECENT_BRANCH_BY_REPOSITORY">
+      <map>
+        <entry key="$PROJECT_DIR$" value="master" />
+      </map>
+    </option>
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
   </component>
   <component name="MarkdownSettingsMigration">
@@ -59,17 +58,20 @@
     <key name="CreateTestDialog.Recents.Supers">
       <recent name="com.keyin.qap4.abstractclasses.Shape" />
     </key>
+    <key name="MoveClassesOrPackagesDialog.RECENTS_KEY">
+      <recent name="com.keyin.qap4" />
+    </key>
     <key name="CreateTestDialog.RecentsKey">
       <recent name="com.keyin.qap4.abstractclasses" />
     </key>
   </component>
   <component name="RunManager" selected="Application.Demo">
     <configuration name="Demo" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
-      <option name="MAIN_CLASS_NAME" value="com.keyin.qap4.abstractclasses.Demo" />
+      <option name="MAIN_CLASS_NAME" value="com.keyin.qap4.Demo" />
       <module name="AbstractClasses_Interfaces" />
       <extension name="coverage">
         <pattern>
-          <option name="PATTERN" value="com.keyin.qap4.abstractclasses.*" />
+          <option name="PATTERN" value="com.keyin.qap4.*" />
           <option name="ENABLED" value="true" />
         </pattern>
       </extension>
@@ -82,7 +84,7 @@
       <module name="AbstractClasses_Interfaces" />
       <extension name="coverage">
         <pattern>
-          <option name="PATTERN" value="com.keyin.qap4.abstractclasses.*" />
+          <option name="PATTERN" value="com.keyin.qap4.*" />
           <option name="ENABLED" value="true" />
         </pattern>
       </extension>
@@ -116,6 +118,14 @@
       <option name="presentableId" value="Default" />
       <updated>1668954013151</updated>
     </task>
+    <task id="LOCAL-00001" summary="Flesh of the program--problem 1">
+      <created>1669663082339</created>
+      <option name="number" value="00001" />
+      <option name="presentableId" value="LOCAL-00001" />
+      <option name="project" value="LOCAL" />
+      <updated>1669663082340</updated>
+    </task>
+    <option name="localTasksCounter" value="2" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -129,11 +139,15 @@
       </map>
     </option>
   </component>
+  <component name="VcsManagerConfiguration">
+    <MESSAGE value="Flesh of the program--problem 1" />
+    <option name="LAST_COMMIT_MESSAGE" value="Flesh of the program--problem 1" />
+  </component>
   <component name="XDebuggerManager">
     <breakpoint-manager>
       <breakpoints>
         <line-breakpoint enabled="true" type="java-field">
-          <url>file://$PROJECT_DIR$/src/main/java/com/keyin/qap4/abstractclasses/Shape.java</url>
+          <url>file://$PROJECT_DIR$/src/main/java/com/keyin/qap4/Shape.java</url>
           <line>7</line>
           <properties field="name" class="com.keyin.qap4.abstractclasses.Shape" />
           <option name="timeStamp" value="1" />
