+++ src/main/java/com/keyin/qap4/Scalable.java
@@ -1,0 +1,6 @@
+package com.keyin.qap4;
+//abstract redundant for interface since interface is always abstract
+public interface Scalable {
+//public and abstract not necessary since that's already defined in the interface
+    void scale(int scalingFactor);
+}
